Moves subtitle settings panel to subtitle folder
